trying to add login with facebook

+++ db/migrate/20210329194506_add_uid_to_user.rb
@@ -1,5 +1,5 @@
 class AddUidToUser < ActiveRecord::Migration[6.0]
   def change
-    add_column :users, :uid, :integer
+    add_column :users, :uid, :numeric
   end
 end
